+++ examples/client/getConsensusStatus.ts
@@ -1,0 +1,42 @@
+import { ConsensusStatus, createConcordiumClient } from '@concordium/node-sdk';
+import { credentials } from '@grpc/grpc-js';
+
+import meow from 'meow';
+
+const cli = meow(
+    `
+  Usage
+    $ yarn ts-node <path-to-this-file> [options]
+
+  Options
+    --help,         Displays this message
+    --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
+`,
+    {
+        importMeta: import.meta,
+        flags: {
+            endpoint: {
+                type: 'string',
+                alias: 'e',
+                default: 'localhost:20000',
+            },
+        },
+    }
+);
+
+const [address, port] = cli.flags.endpoint.split(':');
+const client = createConcordiumClient(
+    address,
+    Number(port),
+    credentials.createInsecure()
+);
+
+/**
+ * Retrieves the current consensus status from the node.
+ */
+
+(async () => {
+    const consensusStatus: ConsensusStatus = await client.getConsensusStatus();
+
+    console.dir(consensusStatus, { depth: null, colors: true });
+})();